allocate db rows for image/video representations before creating
This allows creating a unique id for every representation, from which its storage key is derived.
Before this change, the storage key was derived from info about the representation,
instead of its unique id which is more convenient.
After processing, fields only known after everything is done like size and bitrate are written to the db.