web: close gallery even if slide content element hasn't loaded
Previously, the gallery would only close when the transition applied to the image/video element finished.
If the content hadn't loaded yet, that would never happen.
Now the transition is applied to the placeholder image if it happens to still be visible.
In case all else fails, we just pretend the transition happened and call the callback.