web: fix timeline scroll adjustments
Previously, the timeline set window.scrollY to compensate for changes in section heights.
After putting the timeline in its own scroll container, this did not do anything anymore.
Now, the timeline store can signal that container scroll needs to be adjusted, which the
owner of the scroll wrapper element executes.