web: optimize image slide loading
Before this change, image slides were fetched by setting the <img>
elements's src attribute. This caused a few problems:
 - requests are not (necessarily) cancelled when the user clicks away
   and the <img> element is removed
 - all slides are loaded simultaneously, when really the currently
   active (center) one is most important

With this change, the active slide is prioritized, and the ones to the
left and right are only loaded after.
Any running requests are canceled when slides are removed or switch
content.